+++ src/main/java/net/terrafirmainfinity/core/datagen/lang/InfinityLangHandler.java
@@ -6,6 +6,7 @@ import com.tterrag.registrate.providers.RegistrateLangProvider;
 public class InfinityLangHandler extends LangHandler {
     public static void init(RegistrateLangProvider provider) {
         replaceMaterialLang(provider, "andesite_alloy", "Andesite Composite");
+        replaceMaterialLang(provider, "dawnstone", "Dawnsteel");
 
         provider.add("tagprefix.poor_raw_ore", "Poor Raw %s");
         provider.add("tagprefix.rich_raw_ore", "Rich Raw %s");